cyethack/requirements.txt
asgiref==3.11.0
django==5.2.9
djangorestframework==3.16.1
django-filter==22.1
sqlparse==0.5.5
tzdata==2025.3
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1